requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy==2.0.36
pymysql==1.1.1
python-dotenv==1.0.1
akshare>=1.14.0
pandas>=2.2.0
numpy>=1.26.0
apscheduler==3.10.4
cryptography==44.0.0
